[VMWare] Refactor handling of ansible playbooks parameters
Combine dict with all the required data first and only then use it
for running ansible playbooks. It allows us to use optional playbook
vars based on configuration.
Use case: some playbooks depend on "is defined" option state, so,
with current feature we are able to avoid definition of var when we do
not need it to be defined.